==> 20-eda-treino.qmd <==
---
title: "Titanic competition: EDA (só df de treino)"
author: "Fernando Náufel"
date: now
date-format: "DD/MM/YYYY HH:mm"
lang: pt

#bibliography: bibliography.bib
#csl: universidade-do-porto-faculdade-de-engenharia-chicago-pt-crono.csl

format:
  html:
    toc: true  
    toc-depth: 3
    number-depth: 3
    theme: 
      - journal
      - _custom.scss
    link-external-icon: true
    link-external-newwindow: true
    link-external-filter: ^(?:http:|https:)\/\/fnaufel\.github\.io
    df-print: paged
---

{{< include _math.qmd >}}


# Setup

```{r setup}
source('_setup.R')
source('R/functions.R')
```


# Leitura e receita pós-leitura

```{r}
df <- read_csv(
  'dados/train_limpo.csv'
)
```

```{r}
df <- receita_pos_leitura(df) %>% 
  prep(
    verbose = TRUE,
    retain = TRUE,
    log_changes = TRUE,
    strings_as_factors = FALSE,
  ) %>% 
  bake(new_data = NULL) %>% 
  mutate(survived = survived == 1)
```

```{r}
df %>% glimpse()
```

## Sumário

```{r}
df %>% dfSummary() %>% print(method = 'render')
```


# Sobrevivência por classe

```{r}
df %>% 
  group_by(pclass) %>% 
  summarize(n = n(), .groups = 'drop')
```

```{r}
df %>% 
  group_by(pclass, survived) %>% 
  summarize(n = n(), .groups = 'drop')
```

Absolutos:

```{r}
df %>% 
  ggplot(aes(pclass, fill = survived)) +
    geom_bar(position = 'dodge') +
    labs(y = NULL)
```

Proporções:

```{r}
df %>% 
  ggplot(aes(pclass, fill = survived)) +
    geom_bar(position = 'fill') +
    labs(y = NULL)
```


# Sobrevivência por sexo

```{r}
df %>% 
  ggplot(aes(sex, fill = survived)) +
    geom_bar(position = 'fill') +
    labs(y = NULL)
```


# Sobrevivência por sexo e classe

Quantas pessoas havia por classe e sexo?

```{r}
df %>% 
  group_by(pclass, sex) %>% 
  summarize(n = n(), .groups = 'drop')
```


Proporções:

```{r}
df %>% 
  ggplot(aes(sex, fill = survived)) +
    geom_bar(position = 'fill') +
    labs(y = NULL) +
    facet_wrap(~pclass)
```

Absolutos:

```{r}
df %>% 
  ggplot(aes(sex, fill = survived)) +
    geom_bar(position = 'dodge') +
    labs(y = NULL) +
    facet_wrap(~pclass)
```

::: {.callout-important collapse="true"}

## Classe e sexo

Quase todas as mulheres da primeira e da segunda classes sobreviveram!

Só estas morreram:

```{r}
df %>% 
  filter(!survived & sex == 'F' & pclass %in% c(1, 2)) %>% 
  select(pclass, age, title, family, fare) %>% 
  arrange(pclass, desc(age))
```

:::


# Sobrevivência por passagem

```{r}
df %>% 
  ggplot(aes(survived, fare)) +
    geom_boxplot()
```

Só passagens abaixo de $100$:

```{r}
df %>% 
  filter(fare < 100) %>% 
  ggplot(aes(survived, fare)) +
    geom_boxplot()
```


# Sobrevivência por idade

```{r}
df %>% 
  ggplot(aes(age, fill = survived)) +
    geom_histogram(breaks = seq(0, 80, 5)) +
    facet_wrap(~survived, ncol = 1) +
    scale_x_continuous(breaks = seq(0, 80, 5)) +
    labs(y = NULL)
```


# Sobrevivência por número de pais e filhos

```{r}
df %>% 
  count(par_ch, survived)
```

```{r}
df %>% 
  ggplot(aes(par_ch, fill = survived)) +
    geom_bar(position = 'dodge') +
    scale_x_continuous(breaks = 0:6) +
    labs(y = NULL)
```


# Sobrevivência por número de irmãos e cônjuges

```{r}
df %>% 
  count(sib_sp, survived)
```

```{r}
df %>% 
  ggplot(aes(sib_sp, fill = survived)) +
    geom_bar(position = 'dodge') +
    scale_x_continuous(breaks = 0:8) +
    labs(y = NULL)
```


# Títulos e idades

```{r}
df %>% 
  ggplot(aes(fct_reorder(title, -age, .na_rm = TRUE), age)) +
    geom_boxplot() +
    geom_point(alpha = 0.3, size = 1, color = 'blue') +
    scale_y_continuous(breaks = seq(0, 80, 5)) +
    coord_flip() +
    labs(x = NULL)
```

::: {.callout-important collapse="true"}

## Títulos e idades

Várias idades estão faltando e vão precisar ser imputadas. Os títulospodem ajudar.

* Todos os *Masters* são crianças (${}<14$). 
* As *Misses* podem ser de qualquer idade.
* Todos os restantes são adultos.

:::


# Sobrevivência por porto de embarque

```{r}
df %>% 
  count(embarked, survived)
```

```{r}
df %>% 
  ggplot(aes(embarked, fill = survived)) +
    geom_bar(position = 'dodge')
```

::: {.callout-important collapse="true"}

## Porto de embarque importa!

Quem embarcou em *S* teve $2/3$ de probabilidade de morrer.

Quem embarcou nos outros portos teve $1/2$ de probabilidade de morrer.

:::
